wip: move config files
