+++ e2e/config-file/test/cli/append/append-config.js
@@ -1,4 +1,4 @@
-const OutputFileNames = require("../test/output-filenames");
+const OutputFileNames = require("../../output-filenames");
 
 module.exports = {
   inputs: ["./package.json"],
