+++ e2e/config-file/test/cli/replacement/replacement.e2e.spec.ts
@@ -1,16 +1,16 @@
 import { exec } from "child_process";
 import fs from "fs/promises";
 import { promisify } from "util";
-import * as OutputFileNames from "../output-filenames.js";
+import * as OutputFileNames from "../../output-filenames.js";
 
 const execAsync = promisify(exec);
 
 describe("cli", () => {
   beforeEach(async () => {});
 
-  it("should match snapshot when content is excluded", async () => {
-    const configPath = "./.config/exclude-config.js";
-    const outputPath = OutputFileNames.ExcludeConfig;
+  it("should match snapshot when a package is replaced", async () => {
+    const configPath = "./.config/replacement-config.js";
+    const outputPath = OutputFileNames.ReplacementConfig;
 
     await execAsync(`npx generate-license-file -c ${configPath}`);
